tests/wpt/mozilla/tests/mozilla/document_title_nontextchildren.html

<html>
<title>x</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  document.head.firstChild.appendChild(document.createElement("foo"));
  assert_equals(document.title, "x");
});
</script>
</html>
